requirements.txt
absl-py==0.11.0
APScheduler==3.6.3
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
fpdf==1.7.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.10
img2pdf==0.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.6.1
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
pikepdf==1.19.3
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
